requirements.txt
Click==7.0
Flask==2.1.3
infi.clickhouse_orm==2.2.2
